moved ScheduleAddCell to same file as ColorAndDelete, and renamed to reflect the chagne
